refactor: move download ad code into components; remove lib/downloadAd*
